pyproject.toml
[tool.poetry]
name = "dataclass_ext"
version = "0.1.0"
description = "Python dataclass extensions"
authors = ["Roman Koifman <roma.koifman@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"


[tool.poetry.group.aws.dependencies]
boto3 = "^1.34.39"


[tool.poetry.group.mongo.dependencies]
pymongo = "^4.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
